+++ src/core/requests/headers.py
@@ -235,7 +235,7 @@ def check_http_traffic(request):
     # Checks regarding recognition of generic "your ip has been blocked" messages.
     checks.blocked_ip(page)
     # Checks for not declared cookie(s), while server wants to set its own.
-    if menu.options.cookie == None:
+    if menu.options.cookie == None and not menu.options.drop_set_cookie:
       checks.not_declared_cookies(response)
 
   # This is useful when handling exotic HTTP errors (i.e requests for authentication).